+++ src/wai/common/statistics/__init__.py
@@ -1,0 +1,2 @@
+from ._functions import lower_quartile, lower_quartile_sorted, upper_quartile, upper_quartile_sorted, \
+    interquartile_range