+++ CHANGES.rst
@@ -23,3 +23,4 @@ Changelog
 - Added exception InvalidStateError for classes that get into an invalid setup.
 - Added typing module for type-related functionality.
 - Added first, which finds the first element of an iterable to match a predicate.
+- Added statistics package with quartile functions.